Ring Toss Refactoring.
Refactored all of the code related to the ring toss mini game. A lot has changed, added a ring factory and setup class for dynamic ring positioning and instantiating based on the amount of desired rings.

Completely changed the oscillation class of the force bar to be more concise, the original script 'Oscillator' is kept for posterity.

Still more that could be done, but I'm happy for now.