Tilt and Rotate Voice Commands for TiltShrine
Working tilt and move voice commands, yet to be tested end to end, using microphone input however.

There's a bug with the tilting of the arena, the walls and base are not tilting at the same rate(?), not sure why. - VoiceCommands.cs 198 and below probably. Also the axis' are not correct, "left" tilts up for example.